+++ src/sunflare/controller/__init__.py
@@ -1,10 +1,6 @@
-from ..virtual._mixin import AsyncSubscriber, Publisher, SyncSubscriber
 from ._protocols import ControllerProtocol, HasShutdown
 
 __all__ = [
-    "HasShutdown",
     "ControllerProtocol",
-    "Publisher",
-    "SyncSubscriber",
-    "AsyncSubscriber",
+    "HasShutdown",
 ]
